requirements.txt
openai==1.3.0
opencv-python
pillow==10.1.0
requests==2.31.0